we only need one util file; types conflicts with stdlib; ipynb test to match new behavior
