+++ coorx/__init__.py
@@ -1,11 +1,9 @@
-from ._util import DependentTransformError
+from ._util import DependentTransformError, AxisSelectionEmbeddedTransform
 from .base_transform import Transform, InverseTransform
 from .composite import CompositeTransform, SimplifiedCompositeTransform
 from .linear import NullTransform, TTransform, STTransform, AffineTransform, SRT3DTransform, TransposeTransform
 from .nonlinear import LogTransform, PolarTransform
 from .coordinates import Point, PointArray, Vector, VectorArray
-from .util import AxisSelectionEmbeddedTransform
-
 
 __version__ = "1.0.5"
